+++ VisualChatGPTStudioShared/Commands/OptimizeDiffView.cs
@@ -48,7 +48,7 @@ namespace JeffPires.VisualChatGPTStudio.Commands
 
                 CancellationTokenSource = new CancellationTokenSource();
 
-                string result = await ChatGPT.GetResponseAsync(OptionsGeneral, command + PROVIDE_ONLY_CODE_INSTRUCTION, selectedText, OptionsGeneral.StopSequences?.Split(','), CancellationTokenSource.Token);
+                string result = await ChatGPT.GetResponseAsync(OptionsGeneral, command + Constants.PROVIDE_ONLY_CODE_INSTRUCTION, selectedText, OptionsGeneral.StopSequences?.Split(','), CancellationTokenSource.Token);
 
                 result = TextFormat.RemoveCodeTagsFromOpenAIResponses(result.ToString());